Replace database models and remove serializers
We've decided to make the db the format that makes the most sense for
data interrelations, which makes it unsuitable for using django
serializers for the xlogfile.

Since this isn't running in production anywhere and changes to models.py
are effectively "new", this commit wipes the old migrations. It also
comments out a lot of code that depends on the old models.

Current status: while the server can be brought up and run, it will
serve only the static template pages, and there is no way to pull data
into the database.