Add setup for loading achievements data into the database
Achievements are a fairly constant piece of data (e.g. will not have to
be updated while the tournament runs), so it makes sense to provide them
as a fixture that can be loaded into the database with django's loaddata
command.

This is a bit more involved of a process for achievements than trophies
and conducts will be, since I wanted to preserve the aspect of the TNNT
source code being the ultimate source of truth to prevent any desyncs...

Trophies and conducts fixtures coming soon.