Add fixtures for achievements, conducts, and trophies
As noted in last commit, these things don't change over the course of
the tournament, so there isn't much need to do anything complicated with
loading them. Conducts and trophies are even less likely to change than
achievements, so they're just static yaml files.

This also updates the database models to fix conduct.name being too
short for some of the "never kill foo" ones, and adds some safety checks
in the form of uniqueness constraints on the xlogfield/bit pairings for
conducts and achievements. The db now won't let you add a conduct or
achievements that has the same xlogfile representation as an existing field.

I wavered a bit on whether or not to commit achievements.yaml to git or
not. Ended up deciding to, since it'll be easier to get the server up
and running without forcing the person to download a copy of the tnnt
source.