Add native user authorization, login, and clan management
Moving away from the old scoreboard's approach, where "clan management"
took you to a completely different site in which you did clan things.
This approach worked fine if the clan database was the only true
database, and the entire rest of the site was statically generated based
on its contents plus xlogfile contents. But that's not the type of site
we're trying to build now. Clans and players are an integral part of
the database, and keeping it an arm's length away didn't make sense.

The result: the clan management and login system, rewritten from the
ground up in Django. I've tried to match functionality if not appearance
with the old system. Namely: anyone can make a clan, becoming the admin
of that clan; they can then invite people to the clan and manage
membership by bestowing admin rights, making and rescinding invites, and
kicking members out if necessary. The frontend styling isn't top-notch,
but the data is there, it's functional, and it doesn't look like a
complete wreck, so this is no time to be perfectionist about it.

This also tweaks a couple other things; I added a TODO here and there
when I noticed something to be done, and as an experiment, <a> elements
have their underline restored to make their link-ness more obvious.