Add native user authorization, login, and clan management  (#9)
* Fix freenode link

* Add clan message, clan admin, and clan invites to models

* Add small script to automate work of un-migrating and re-migrating

Which is a common thing to want to do when the project is new and you
don't have database compatibility with previous versions to worry about.

Running this will wipe all migrations and create a new 0001_initial
migration, based off the current state of models.py.

* Add native user authorization, login, and clan management

Moving away from the old scoreboard's approach, where "clan management"
took you to a completely different site in which you did clan things.
This approach worked fine if the clan database was the only true
database, and the entire rest of the site was statically generated based
on its contents plus xlogfile contents. But that's not the type of site
we're trying to build now. Clans and players are an integral part of
the database, and keeping it an arm's length away didn't make sense.

The result: the clan management and login system, rewritten from the
ground up in Django. I've tried to match functionality if not appearance
with the old system. Namely: anyone can make a clan, becoming the admin
of that clan; they can then invite people to the clan and manage
membership by bestowing admin rights, making and rescinding invites, and
kicking members out if necessary. The frontend styling isn't top-notch,
but the data is there, it's functional, and it doesn't look like a
complete wreck, so this is no time to be perfectionist about it.

This also tweaks a couple other things; I added a TODO here and there
when I noticed something to be done, and as an experiment, <a> elements
have their underline restored to make their link-ness more obvious.

* Updates to clan management: clan size cap, fix a couple links

* Enshrine some of aoei's feedback in FUTURE TODO comments

Things that could probably be improved, but are not a priority for the
2021 tournament.

* Ignore migration files and remove existing one from version control

TNNT doesn't have long-term persistence; it doesn't need to run all
year and may well have different database models next year. Thus we're
not particularly interested in tracking database migrations or
constraining models.py changes to be backwards compatible with previous
database migrations.

* Implement clan freeze

I didn't realize at the time that this is handled on the backend, but of
course it is. This is a configurable time (CLAN_FREEZE_TIME) after which
joining, inviting, and creating clans cannot be done. By default and
tradition, it's Nov 10 at midnight.

* Fix a few clan management bugs

1. leave_clan didn't actually check that the player was in a clan
2. making admin and kicking were still referencing an "otherplayer"
   variable that didn't exist in those functions and would have crashed
3. aoei suggested not passing raw unvalidated POST data through to
   errmsg
4. some error messages still ended with punctuation; de facto standard
   for these is that they shouldn't

* Use form.cleaned_data[] instead of request.POST for the two forms

* Sort player lists on clan management page

Clan members by admin status first, then alphabetized. Invites just
alphabetized.

* Add proper logging

Rather than a homegrown impossible() function that logged only to the
console, or print statements, this commit logs all clan management
events properly. I was noticing that I was trying to classify different
prints as different things - logging's multiple levels made sense for
this.

Most "impossible" cases - the majority of them actually possible if
clicking links in an outdated page or sending POST requests manually -
generate warnings. Normal clan events going through generate info
messages.

* remove stray print statement in hdf auth backend