WIP commit for revamping aggregate to work with queries
Existing process iterated over Games. This doesn't exactly work, since
it means stale data can be left in a clan after a member who helped it
earn its laurels departs. (Also, I have doubts on the efficiency of
calling save() for every Clan and Player once per each of the tens of
thousands of games in the database.)

Instead, clan and player data is now updated with queries. This adds
some of the associated legwork - I implemented a method for getting
unique deaths, but left streaks unimplemented for now. (And trophies.)

A benefit of this approach is that if all the Players get their
aggregate leaderboard stats computed first, the Clans can mostly just
aggregate off those stats, without the need to reach back into the set
of Games to compute anything.