Restore dumplog_fmt in Source and add Game.get_dumplog()
Another late requirement - I realized we need to link to dumplogs. After
some agonized messing around in the templates with custom logic based on
the server being "hdf" or "hfe" etc, I figured out how to do it cleanly
on Game in a way that can be called from the template.

Sources now have a required dumplog_fmt field, which uses a few custom
format specifiers I made up on the fly. When [Game object].get_dumplog()
is called, it returns the URL to that game's dumplog.