Add list-of-players page at /players
The first of several pages to come. The players list is relatively
straightforward, being just a simple table.

Changed some things from the old site - removed the numbers next to each
player, since this isn't a unified scoreboard anymore. Also removed the
score and max dlvl columns - the score because we have no score
computation, and the max dlvl because why was that even here?? Replaced
it with the maximum amount of conducts in an ascension. (Kind of
arbitrary. We're not going to go an add all the leaderboard fields
here.)