Aggregate with queries (#11)
* WIP commit for revamping aggregate to work with queries

Existing process iterated over Games. This doesn't exactly work, since
it means stale data can be left in a clan after a member who helped it
earn its laurels departs. (Also, I have doubts on the efficiency of
calling save() for every Clan and Player once per each of the tens of
thousands of games in the database.)

Instead, clan and player data is now updated with queries. This adds
some of the associated legwork - I implemented a method for getting
unique deaths, but left streaks unimplemented for now. (And trophies.)

A benefit of this approach is that if all the Players get their
aggregate leaderboard stats computed first, the Clans can mostly just
aggregate off those stats, without the need to reach back into the set
of Games to compute anything.

* Add games_scummed to LeaderboardBaseFields

Forgot that we track this for a trophy, and it's used in enough places
that awkwardly querying games everywhere it's needed is inefficient, so
some denormalization is in order (like for the rest of the aggregated
fields).

We're not planning on having an actual leaderboard for this, though...
uh... right?

* Restore dumplog_fmt in Source and add Game.get_dumplog()

Another late requirement - I realized we need to link to dumplogs. After
some agonized messing around in the templates with custom logic based on
the server being "hdf" or "hfe" etc, I figured out how to do it cleanly
on Game in a way that can be called from the template.

Sources now have a required dumplog_fmt field, which uses a few custom
format specifiers I made up on the fly. When [Game object].get_dumplog()
is called, it returns the URL to that game's dumplog.

* Fix: Gracefully handle "orphaned user" in clan management

I would be surprised if this comes up during the actual tournament, but
it happened during testing: I wiped the db to re-apply migrations, but
this didn't wipe Django's built in user table. So I was still able to
log in and access my clan management page, which promptly errored when I
tried to do anything with it.

This adds a fallback where if a user logs in and doesn't have an id
match on Player, it will look for a matching name and link the user and
Player together if found. If there's a logged in user who doesn't match
any Player's name in the database at all, then no. Fail. This is the
first case where we have to return an actual error to the user. (Again,
I don't see this case happening in production because we wouldn't be
casually wiping the Player table.)

* Add list-of-players page at /players

The first of several pages to come. The players list is relatively
straightforward, being just a simple table.

Changed some things from the old site - removed the numbers next to each
player, since this isn't a unified scoreboard anymore. Also removed the
score and max dlvl columns - the score because we have no score
computation, and the max dlvl because why was that even here?? Replaced
it with the maximum amount of conducts in an ascension. (Kind of
arbitrary. We're not going to go an add all the leaderboard fields
here.)

* Add single-player page at /player/<name>

This is the view on a single player, largely the same as the old
scoreboard. It's not complete yet - it lacks trophies and streaks which
the backend isn't computing yet. Apart from that, everything else is
complete.

The biggest change here, apart from score removal, is the achievements
now display in a table rather than a grid. The table rows light up gold,
same as they used to, when the achievement is earned.

* Add partial leaderboards page at /leaderboards

The main new big feature of 2021 is leaderboards. Given their
prominence, they get a whole page linked from the top (replacing
Trophies... but we still have trophies, so we'll need to figure out some
way to navigate to those.)

All leaderboards are loaded on the page, but only one is shown at a
time (because they could get quite long with all the players/clans).
They're currently changed via a combo box with a bit of JS bound to it.
I kind of want to add at least left/right arrow key navigation as well.

As of this commit, only 3 leaderboards are functional, but others are
expected to follow easily along the same pattern.

Also adds a shorthand for querying only ascended players via a Manager
class (Player.ascended.*).

Co-authored-by: copperwater <aosdict@gmail.com>