Revamp achievements display
The theme of "dark blue with white text which turns to goldenrod with
black text" was a holdover from having these all in a grid and it didn't
look very good at all. Instead, turn achievements into a three-column
table (up from 2), where the first column contains a green check or red
X signifying whether it's achieved. This is a bit of a departure from
the site's regular aesthetic of blue/gold/white, but it looks a good
deal better than either gold checks and white Xs, and much more
immediately obvious what is still unachieved than it would be with just
white for both.