Add a command to wipe the database
This may not see production use, but it is certainly useful in dev to be
able to reset everything to a blank slate without the sledgehammer
approach of reset_db.sh tearing the whole database down and rebuilding
it.

Future extension might be to make this take arguments to only wipe
certain tables (for instance, when testing with a changed test xlogfile,
only Game really needs to be wiped).