Implement trophy computation
This happens in aggregation, after all other fields have been aggregated
(which saves a bit of work since e.g. All Achievements can then just
check a player's or clan's unique achievements rather than going to the
database for them).

This went through several iterations, but I'm happiest with this one.
I primarily used sets to compute role-race-gender-align combinations and
compare them to hardcoded lists of what the trophy requires. (Also, in
retrospect, I don't think this would really have worked as uniformly
coded as an additional field on Trophy like the initial envisioning
was when we were still building the database. Maybe it could be made to
work, but now there isn't much of a need.)

Trophies are not removed from Players except for Never Scum a Game
(for which which its removal is handled), but they are removed from
Clans every time aggregation is run, because someone may have left the
clan so they no longer meet the criteria for that trophy.