Move ratio to LeaderboardBaseFields and add check for 0 games
This a) fixes a bug where it was undisplayable on the single player/clan
view if the player/clan had no games, and b) allows clans' ascension
ratio to be shown on their page. Showing the ascension ratio of a clan
may go against TNNT philosophy somewhat in that a weaker player on a
clan can drag that down, but it's not a leaderboard and was on the old
scoreboard, so we're keeping it here.