Merge single player and single clan into one view and template
The new functionality in this commit is to add the /clan/foo route to
the webserver. I started to make a new template for a single clan then
figured it was a lot of duplicate work since the single player and
single clan page are so similar. With not a ton of extra effort, it was
possible to make the same view work for both.

The main differences are that the clan view shows a list of members,
and does not show streaks.