Fix bottom stats on main page
This contained summary tables of the highest-tournament-scoring players
and clans (now no longer needed without tournament score) and lists of
the last 10 ascensions and last 10 games in the tournament, which are
still good to have. In order to get those, I had to take a similar
approach as the single player/clan view -- so similar that I ended up
merging the tables in that view into one template, and reusing it here.

As a result, that system was overhauled somewhat, and it became
necessary to have a semi-standard format of "list of game dicts" passed
through both the index and single player/clan views. The latter now uses
only 2 queries to get its asc/non-asc games data - though it then executes a
query per game to get its conducts.