Resolve or defer all existing TODOs, remove <!-- html comments
Most TODOs are deferred with marks of "future" or "post 2021". Some are
resolved.

In particular, the clan management page has a note to the effect that
clan membership changes won't be reflected immediately on
leaderboards/trophies, since the de facto decision is not to trigger an
aggregation every time someone's membership changes (probably wise).

Also in this commit:
- No HTML comments are important to send to the browser, so convert them
  all into Django template comments.
- Add a player's longest streak to the Players list.