Fix: clan game aggregation included player's null games
Essentially, if a clan had a player in it with no wins (and therefore
null for all the *_asc leaderboard fields), the query for "select the
min/max of the *_asc field over all the members in this clan" would
produce a null. Thus producing a clan in the database which had wins > 0
but nulls for the actual games.

Dodge this problem by filtering these aggregation queries on clan
players that have at least 1 win (or in the mostachgame case, at least 1
game). Also, now only run the aggregations (besides mostachgame) on
clans that have at least 1 win among them.