Fix: clan dumplog links were formatted with clan's name
Of course, the dumplog link corresponds to the player and not the game.
The query did not have all the information it needed to get the player's
name when the game on the leaderboard just belongs to a clan, so, sigh,
this needs to add another set of joins in the query so that information
exists.

Also delete a bit of test code that got committed somehow.