Fix: use subsets rather than set equality for Great/Lesser Foo
Set equality only works in the presence of a faulty assumption: the
listed roles or race/align combinations required for the trophy are not
necessarily the whole set of achievable ones in NetHack. In this case, a
player had the four human ascensions required for Great Human - and also
a human ascension of a different role, making the set == return false.